+++ mplaltair/tests/test_convert.py
@@ -27,7 +27,6 @@ def test_encoding_not_provided():  # TODO: move to the parse_chart tests
     chart_spec = alt.Chart(df).mark_point()
     with pytest.raises(ValueError):
         chart = ChartMetadata(chart_spec)
-        # _convert(chart)
 
 def test_invalid_encodings():
     chart_spec = alt.Chart(df).encode(x2='quant').mark_point()
@@ -39,7 +38,6 @@ def test_invalid_encodings():
 def test_invalid_temporal():  # TODO: move to parse_chart tests???
     chart = alt.Chart(df).mark_point().encode(alt.X('quant:T'))
     ChartMetadata(chart)
-    # _convert(chart)
 
 @pytest.mark.parametrize('channel', ['quant', 'ord', 'nom'])
 def test_convert_x_success(channel):
@@ -58,7 +56,6 @@ def test_convert_x_success_temporal(column):
 def test_convert_x_fail():
     with pytest.raises(KeyError):
         chart_spec = ChartMetadata(alt.Chart(df).encode(x='b:N').mark_point())
-        _convert(chart_spec)
 
 @pytest.mark.parametrize('channel', ['quant', 'ord', 'nom'])
 def test_convert_y_success(channel):
@@ -75,7 +72,6 @@ def test_convert_y_success_temporal(column):
 def test_convert_y_fail():
     with pytest.raises(KeyError):
         chart_spec = ChartMetadata(alt.Chart(df).encode(y='b:N').mark_point())
-        _convert(chart_spec)
 
 @pytest.mark.xfail(raises=ValueError, reason="It doesn't make sense to have x2 and y2 on scatter plots")
 def test_quantitative_x2_y2():
@@ -105,10 +101,9 @@ def test_convert_color_success_temporal(column):
     mapping = _convert(chart)
     assert list(mapping['c']) == list(mdates.date2num(df[column].values))
 
-def test_convert_color_fail():  # TODO: What is this covering?
+def test_convert_color_fail():
     with pytest.raises(KeyError):
         chart_spec = ChartMetadata(alt.Chart(df).encode(color='b:N').mark_point())
-        _convert(chart_spec)
 
 @pytest.mark.parametrize('channel,type', [('quant', 'Q'), ('ord', 'O')])
 def test_convert_fill(channel, type):
@@ -128,10 +123,9 @@ def test_convert_fill_success_temporal(column):
     assert list(mapping['c']) == list(mdates.date2num(df[column].values))
 
 
-def test_convert_fill_fail():  # TODO: what is this covering?
+def test_convert_fill_fail():
     with pytest.raises(KeyError):
         chart_spec = ChartMetadata(alt.Chart(df).encode(fill='b:N').mark_point())
-        _convert(chart_spec)
 
 @pytest.mark.xfail(raises=NotImplementedError, reason="The marker argument in scatter() cannot take arrays")
 def test_quantitative_shape():
@@ -141,13 +135,12 @@ def test_quantitative_shape():
 @pytest.mark.xfail(raises=NotImplementedError, reason="The marker argument in scatter() cannot take arrays")
 @pytest.mark.parametrize("column", ["years", "months", "days", "hrs", "combination"])
 def test_convert_shape_fail_temporal(column):
-    chart = ChartMetadata(alt.Chart(df).mark_point().encode(alt.Shape(column)))
-    mapping = _convert(chart)
+    chart = alt.Chart(df).mark_point().encode(alt.Shape(column))
+    convert(chart)
 
 @pytest.mark.xfail(raises=NotImplementedError, reason="Merge: the dtype for opacity isn't assumed to be quantitative")
 def test_quantitative_opacity_value():
     chart = ChartMetadata(alt.Chart(df_quant).mark_point().encode(opacity=alt.value(.5)))
-    mapping = _convert(chart)
 
 @pytest.mark.xfail(raises=NotImplementedError, reason="The alpha argument in scatter() cannot take arrays")
 def test_quantitative_opacity_array():
@@ -174,7 +167,6 @@ def test_convert_size_success_nominal():
 def test_convert_size_fail():
     with pytest.raises(KeyError):
         chart_spec = ChartMetadata(alt.Chart(df).encode(size='b:N').mark_point())
-        _convert(chart_spec)
 
 @pytest.mark.xfail(raises=NotImplementedError, reason="Dates would need to be normalized for the size.")
 @pytest.mark.parametrize("column", ["years", "months", "days", "hrs", "combination"])
@@ -201,12 +193,11 @@ def test_convert_stroke_fail_temporal(column):
 def test_quantitative_x_count_y():
     df_count = pd.DataFrame({"a": [1, 1, 2, 3, 5], "b": [1.4, 1.4, 2.9, 3.18, 5.3]})
     chart = ChartMetadata(alt.Chart(df_count).mark_point().encode(alt.X('a'), alt.Y('count()')))
-    mapping = _convert(chart)
+
 
 @pytest.mark.xfail(raises=NotImplementedError, reason="specifying timeUnit is not supported yet")
 def test_timeUnit():
     chart = ChartMetadata(alt.Chart(df).mark_point().encode(alt.X('date(combination)')))
-    _convert(chart)
 
 # Plots
 
@@ -279,7 +270,6 @@ class TestLines(object):
         )
         fig, _ = convert(chart)
         return fig
-        # plt.show()
 
     @pytest.mark.mpl_image_compare(baseline_dir='baseline_images/test_convert')
     @pytest.mark.parametrize('o', ['d:Q', 'c:O', 'dates:T'])
@@ -291,7 +281,6 @@ class TestLines(object):
         )
         fig, ax = convert(chart)
         return fig
-        # plt.show()
 
     @pytest.mark.mpl_image_compare(baseline_dir='baseline_images/test_convert')
     @pytest.mark.parametrize('c,o', [('d:Q', 'd:Q'), ('c:N', 'c:O'), ('dates:T', 'dates:T')])
